+++ RELEASE_NOTES.md
@@ -1,5 +1,6 @@
 ## New in 0.1.2
 * added string-starts-with assertion
+* addes string-ends-with assertion
 
 ## New in 0.1.1
 * added support for IReadOnlyCollection